Move taxonomy to metagenomics
